Ajout modérateur admin + code stats

+++ Controleur/admin_delete.php
@@ -1,22 +1,48 @@
 <?php
 include "../Modele/Connexion.php";
 session_start();
-$reponse = $BDD->query('SELECT IdMaison FROM maison WHERE IdUtilisateur="'.$_POST['search'].'"');
-$Id = 0;
-while ($donnees = $reponse->fetch()) {
-    $Id = $donnees['IdMaison'];
+
+if (isset($_POST["choix"])) {
+    $choix = htmlspecialchars($_POST['choix']);
+} else {
+    $choix = 0;
 }
 
-$reponse2 = $BDD->query('SELECT IdPiece FROM piece WHERE IdMaison = "' . $Id . '"');
-$Id2 = 0;
-while ($donnees2 = $reponse2->fetch()) {
-    $Id2 = $donnees2['IdPiece'];
-    $req1 = $BDD->exec("DELETE FROM capteur WHERE IdPiece = '" . $Id2 . "'");
+if ($choix=="supprimer"){
+    $reponse = $BDD->query('SELECT IdMaison FROM maison WHERE IdUtilisateur="'.$_POST['search'].'"');
+    $Id = 0;
+    while ($donnees = $reponse->fetch()) {
+        $Id = $donnees['IdMaison'];
+    }
+
+    $reponse2 = $BDD->query('SELECT IdPiece FROM piece WHERE IdMaison = "' . $Id . '"');
+    $Id2 = 0;
+    while ($donnees2 = $reponse2->fetch()) {
+        $Id2 = $donnees2['IdPiece'];
+        $req1 = $BDD->exec("DELETE FROM capteur WHERE IdPiece = '" . $Id2 . "'");
+    }
+
+
+    $req1 = $BDD->exec("DELETE FROM piece WHERE IdMaison = '" . $Id . "'");
+    $req = $BDD->exec('DELETE FROM maison WHERE IdUtilisateur="'.$_POST['search'].'"');
+    $req2 = $BDD->exec('DELETE FROM utilisateur WHERE IdUtilisateur="'.$_POST['search'].'"');
+
+
+    header('location:../Vue/Admin.php');
 }
 
+if ($choix=="donner"){
+    if(isset($_POST["search"])) {
+        $id = $_POST["search"];
+        $reponse = $BDD->exec('UPDATE utilisateur SET Admin = 1 WHERE IdUtilisateur="' . $id . '"');
+    }
+    header('location:../Vue/Admin.php');
+}
 
-$req1 = $BDD->exec("DELETE FROM piece WHERE IdMaison = '" . $Id . "'");
-$req = $BDD->exec('DELETE FROM maison WHERE IdUtilisateur="'.$_POST['search'].'"');
-$req2 = $BDD->exec('DELETE FROM utilisateur WHERE IdUtilisateur="'.$_POST['search'].'"');
-
-header('location:../Vue/Admin.php');
+if ($choix=="retirer"){
+    if(isset($_POST["search"])) {
+        $id = $_POST["search"];
+        $reponse = $BDD->exec('UPDATE utilisateur SET Admin = 0 WHERE IdUtilisateur="' . $id . '"');
+    }
+    header('location:../Vue/Admin.php');
+}
